track6-project/flask_service/requirements.txt
Flask
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-prometheus
prometheus_client
opentelemetry-instrumentation-flask             
